* int size replaced by unsigned size in TokenArray   to keep compatibility with size_t in loops

+++ src/eval.c
@@ -8,7 +8,7 @@ double eval(TokenArray rpn_expr)
     stack_t* stack = create_stack();
     char* endptr; /* strtod stuff */
 
-    for (int i = 0; i < rpn_expr.size; i++) {
+    for (size_t i = 0; i < rpn_expr.size; i++) {
         if (is_double(rpn_expr.array[i])) { /* check if double */
             push(stack, rpn_expr.array[i]);
         } else {
